+++ src/app/forms/selects-page/selects-page.component.ts
@@ -15,6 +15,7 @@ export function forbiddenNameValidator(nameRe: RegExp): ValidatorFn {
 @Component({
   selector: 'app-selects-page',
   templateUrl: './selects-page.component.html',
+  styleUrls: ['./selects-page.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush,
   standalone: false
 })